Still need to rework damage and hp to be doubles instead of damage :(

+++ src/grid/Tile.java
@@ -184,7 +184,7 @@ public class Tile {
 		trapRefs.add(point);
 	}
 
-	public void tick(long milli) {
+	public void tick(double milli) {
 		
 	}
